DR-471 custodians and stewards read access on studies (#187)
* first pass

* adding json files for testing ingest

* update integration test around custodian access

* cleanup connectedOperations

* rename flappy -> eventual

* pr comments

* failed travis run fix + pr comments

+++ ops/db-connect.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 
 : ${DB:?}
+: ${ENVIRONMENT:?}
+: ${SUFFIX:-$ENVIRONMENT}
 
-PW=$( vault read -format=json secret/dsde/datarepo/integration/api-secrets-integration.json | \
+VAULT_PATH="secret/dsde/datarepo/${ENVIRONMENT}/api-secrets-${SUFFIX}.json"
+echo $VAULT_PATH
+PW=$( vault read -format=json $VAULT_PATH | \
       jq -r .data.datarepoPassword )
 
 kubectl --namespace data-repo run psql -it --serviceaccount=jade-sa --restart=Never --rm --image postgres:9.6 -- \
